backend/requirements.txt
asgiref==3.8.1
astroid==2.4.2
certifi==2024.2.2
colorama==0.4.4
distlib==0.3.8
Django==4.2.11
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.8.6
psycopg2-binary==2.9.3
PyJWT==2.10.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.25.3
wrapt==1.12.1
